Move the empty planNode to its own source file.
