+++ sql/plan.go
@@ -503,45 +503,3 @@ var _ planNode = &dropTableNode{}
 var _ planNode = &dropIndexNode{}
 var _ planNode = &alterTableNode{}
 
-// emptyNode is a planNode with no columns and either no rows (default) or a single row with empty
-// results (if results is initialized to true). The former is used for nodes that have no results
-// (e.g. a table for which the filtering condition has a contradiction), the latter is used by
-// select statements that have no table or where we detect the filtering condition throws away all
-// results.
-type emptyNode struct {
-	results bool
-}
-
-func (*emptyNode) Columns() []ResultColumn { return nil }
-func (*emptyNode) Ordering() orderingInfo  { return orderingInfo{} }
-func (*emptyNode) Values() parser.DTuple   { return nil }
-func (*emptyNode) Err() error              { return nil }
-
-func (*emptyNode) ExplainPlan(_ bool) (name, description string, children []planNode) {
-	return "empty", "-", nil
-}
-
-func (e *emptyNode) ExplainTypes(_ func(string, string)) {}
-
-func (*emptyNode) MarkDebug(_ explainMode) {}
-
-func (*emptyNode) DebugValues() debugValues {
-	return debugValues{
-		rowIdx: 0,
-		key:    "",
-		value:  parser.DNull.String(),
-		output: debugValueRow,
-	}
-}
-
-func (e *emptyNode) Start() error {
-	return nil
-}
-
-func (e *emptyNode) Next() bool {
-	r := e.results
-	e.results = false
-	return r
-}
-
-func (*emptyNode) SetLimitHint(_ int64, _ bool) {}
